requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bitarray==1.2.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
cytoolz==0.11.2
daphne==3.0.2
dateparser==1.1.0
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fake-useragent==0.1.11
frozenlist==1.3.0
gunicorn==20.1.0
hexbytes==0.2.2
hiredis==2.0.0
httmock==1.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.1
incremental==21.3.0
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
kombu==5.2.3
lru-dict==1.1.7
Markdown==3.3.6
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-binance==1.0.15
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
requests==2.27.1
rlp==2.0.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toolz==0.11.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.27.0
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
